% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{create_function_network}
\alias{create_function_network}
\title{Create Function Network}
\usage{
create_function_network(
  source,
  type = c("files", "objects"),
  env = parent.frame()
)
}
\arguments{
\item{source}{Either a path to R files or a character vector of object names}

\item{type}{One of "files" or "objects" indicating the source type}

\item{env}{Environment to look for objects (default: parent.frame())}
}
\value{
A visNetwork object representing function dependencies
}
\description{
Creates a network visualisation of function dependencies from R files or R objects
}
