% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{create_object_network}
\alias{create_object_network}
\title{Create Object Network}
\usage{
create_object_network(
  source,
  type = c("files", "objects"),
  object_types = c("functions", "data.frames", "all"),
  env = parent.frame()
)
}
\arguments{
\item{source}{Source of objects (files or environment)}

\item{type}{Type of analysis ("files", "objects")}

\item{object_types}{Types of objects to track ("functions", "data.frames", "all")}

\item{env}{Environment to look in}
}
\description{
Create Object Network
}
